\name{dpeak-internal}
\title{Internal dpeak objects}
\alias{.bg_SET}
\alias{.cloglik}
\alias{.constructExtRead}
\alias{.deconCore}
\alias{.deconInit}
\alias{.deconWrapper}
\alias{.dlength}
\alias{.exportBED}
\alias{.exportGFF}
\alias{.exportTXT}
\alias{.ff_bg_PET}
\alias{.ff_bg_PET_iter}
\alias{.ff_bg_SET_iter}
\alias{.ff_ma}
\alias{.ff_score}
\alias{.generateFragment}
\alias{.loglik}
\alias{.plotFit}
\alias{.plotGOF}
\alias{.stackFragment}
\alias{.selectModel}
\description{Internal dpeak objects.}
\details{These are not to be called by the user.}
\keyword{internal}
