\name{exampleData}
\docType{data}
\alias{exampleData}
\title{E. coli ChIP-seq Dataset}
\description{
  This is an example E. coli ChIP-seq dataset.
}
\usage{ data(exampleData) }
\format{
  \code{DpeakData} class object containing aligned reads for a peak.
}
\examples{
\dontrun{
data(exampleData)
exampleData
}
}
\keyword{datasets}
